+++ packages/apptive_grid_theme/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.3.1
 * Fix Lints for Flutter 3.7.0
+* Add Colors and Icons for Kanban, Calendar and Maps
 
 ## 0.3.0
 * Update Input Appearance